requirements-arm.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
attrs=24.3.0=py312hca03da5_0
beautifulsoup4=4.13.3=pypi_0
brotli-python=1.0.9=py312h313beb8_9
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.2.25=hca03da5_0
cachetools=5.5.2=pypi_0
certifi=2025.1.31=py312hca03da5_0
charset-normalizer=3.4.1=pypi_0
expat=2.6.4=h313beb8_0
google=3.0.0=pypi_0
google-api-core=2.24.1=pypi_0
google-api-python-client=2.163.0=pypi_0
google-auth=2.38.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-auth-oauthlib=1.2.1=pypi_0
googleapis-common-protos=1.69.1=pypi_0
h11=0.14.0=py312hca03da5_0
httplib2=0.22.0=pypi_0
icalendar=4.0.7=py_0
idna=3.7=py312hca03da5_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
ncurses=6.4=h313beb8_0
oauth2client=4.1.3=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.16=h02f6b3c_0
outcome=1.3.0=py312hca03da5_0
packaging=24.2=pypi_0
pip=25.0=py312hca03da5_0
proto-plus=1.26.0=pypi_0
protobuf=5.29.3=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.9=h99e199e_0
python-dateutil=2.9.0post0=py312hca03da5_2
python-dotenv=1.0.1=pypi_0
pytz=2024.1=py312hca03da5_0
readline=8.2=h1a28f6b_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
selenium=4.24.0=py312hf0e4da2_0
setuptools=75.8.0=py312hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py312hca03da5_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h80987f9_0
tk=8.6.14=h6ba3021_0
trio=0.26.2=py312hca03da5_0
trio-websocket=0.12.2=py312hca03da5_0
typing_extensions=4.12.2=py312hca03da5_0
tzdata=2025a=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=2.3.0=py312hca03da5_0
webdriver-manager=4.0.2=pypi_0
websocket-client=1.8.0=py312hca03da5_0
wheel=0.45.1=py312hca03da5_0
wsproto=1.2.0=py312hca03da5_0
xz=5.6.4=h80987f9_1
zlib=1.2.13=h18a0788_1
